Week14_ChatGPT_API_Adv_GenAi/GeminiAPI/bio_explain_fastapi/requirements.txt
fastapi
uvicorn
google-genai
jinja2
markdown
